Simplify ToString function for types (avoid unnecessary casting)

+++ utils/utils.go
@@ -26,7 +26,7 @@ func ToString(value interface{}) string {
 	case int32:
 		return strconv.FormatInt(int64(value), 10)
 	case int64:
-		return strconv.FormatInt(int64(value), 10)
+		return strconv.FormatInt(value, 10)
 	case uint8:
 		return strconv.FormatUint(uint64(value), 10)
 	case uint16:
@@ -34,11 +34,11 @@ func ToString(value interface{}) string {
 	case uint32:
 		return strconv.FormatUint(uint64(value), 10)
 	case uint64:
-		return strconv.FormatUint(uint64(value), 10)
+		return strconv.FormatUint(value, 10)
 	case float32:
 		return strconv.FormatFloat(float64(value), 'g', -1, 64)
 	case float64:
-		return strconv.FormatFloat(float64(value), 'g', -1, 64)
+		return strconv.FormatFloat(value, 'g', -1, 64)
 	case bool:
 		return strconv.FormatBool(value)
 	default:
